Move GCE discovery code to command/agent/config_gce.go
